+++ src/main/java/subway/controller/StationController.java
@@ -1,7 +1,10 @@
-package subway;
+package subway.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.controller.dto.request.StationRequest;
+import subway.controller.dto.response.StationResponse;
+import subway.service.StationService;
 
 import java.net.URI;
 import java.util.List;
